+++ lightly_studio/tests/api/routes/api/test_export.py
@@ -10,7 +10,10 @@ from sqlmodel import Session
 from lightly_studio.api.routes.api.status import (
     HTTP_STATUS_OK,
 )
-from lightly_studio.models.annotation.annotation_base import AnnotationCreate, AnnotationType
+from lightly_studio.models.annotation.annotation_base import (
+    AnnotationCreate,
+    AnnotationType,
+)
 from lightly_studio.resolvers import annotation_resolver, tag_resolver
 from tests.helpers_resolvers import (
     ImageStub,