+++ lightly_studio_view/src/routes/datasets/[dataset_id]/[collection_type]/[collection_id]/videos/+page.svelte
@@ -136,7 +136,12 @@
                     dataSampleName={items[index].file_name}
                 >
                     {#snippet item()}
-                        <VideoItem video={items[index]} size={sampleSize} {index} />
+                        <VideoItem
+                            video={items[index]}
+                            size={sampleSize}
+                            {index}
+                            showCaption={true}
+                        />
                     {/snippet}
                 </SampleGridItem>
             {/if}